+++ server/src/main/java/org/jetbrains/bsp/bazel/server/BazelBspServer.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.bsp.bazel.server
 
 import ch.epfl.scala.bsp4j.BuildClient
+import ch.epfl.scala.bsp4j.TextDocumentIdentifier
 import org.eclipse.lsp4j.jsonrpc.Launcher
 import org.jetbrains.bsp.bazel.bazelrunner.BazelRunner
 import org.jetbrains.bsp.bazel.logger.BspClientLogger
@@ -14,6 +15,7 @@ import org.jetbrains.bsp.bazel.server.sync.ExecuteService
 import org.jetbrains.bsp.bazel.server.sync.ProjectSyncService
 import org.jetbrains.bsp.bazel.workspacecontext.WorkspaceContextProvider
 import java.nio.file.Path
+import java.util.concurrent.ConcurrentHashMap
 
 class BazelBspServer(
     bspInfo: BspInfo, workspaceContextProvider: WorkspaceContextProvider, val workspaceRoot: Path
@@ -23,10 +25,11 @@ class BazelBspServer(
     private val bspServerApi: BspServerApi
     private val bspClientLogger: BspClientLogger = BspClientLogger()
     private val bspClientTestNotifier: BspClientTestNotifier = BspClientTestNotifier()
+    private val bspState: Map<String, Set<TextDocumentIdentifier>> = ConcurrentHashMap()
 
     init {
         bazelRunner = BazelRunner.of(workspaceContextProvider, this.bspClientLogger, workspaceRoot)
-        compilationManager = BazelBspCompilationManager(bazelRunner)
+        compilationManager = BazelBspCompilationManager(bazelRunner, bspState)
         bspServerApi = BspServerApi { bspServerData(bspInfo, workspaceContextProvider) }
     }
 
@@ -51,7 +54,8 @@ class BazelBspServer(
             serverContainer.projectProvider,
             bazelRunner,
             workspaceContextProvider,
-            bspClientTestNotifier
+            bspClientTestNotifier,
+            bspState,
         )
         val serverLifetime = BazelBspServerLifetime()
         val bspRequestsRunner = BspRequestsRunner(serverLifetime)