+++ server/src/main/java/org/jetbrains/bsp/bazel/server/bloop/BloopExporter.java
@@ -16,6 +16,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import org.eclipse.lsp4j.jsonrpc.CancelChecker;
@@ -60,7 +61,7 @@ class BloopExporter {
     var bspClientLogger = new BspClientLogger();
     var bspClientTestNotifier = new BspClientTestNotifier();
     var bazelRunner = BazelRunner.of(workspaceContextProvider, bspClientLogger, workspaceRoot);
-    var compilationManager = new BazelBspCompilationManager(bazelRunner);
+    var compilationManager = new BazelBspCompilationManager(bazelRunner, new ConcurrentHashMap());
     var serverContainer =
         ServerContainer.create(
             bspInfo,